@@ -1,10 +1,10 @@
 name: month_picker_dialog
 description: Internationalized dialog for picking a single month from an infinite list of years.
-version: 0.4.0
+version: 0.5.0
 homepage: https://github.com/hmkrivoj/month_picker_dialog
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -13,7 +13,7 @@ dependencies:
     sdk: flutter
   intl: ">=0.1.0<2.0.0"
   #flutter_cupertino_localizations: ^1.0.1
-  rxdart: ^0.26.0
+  rxdart: ^0.27.5
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
