@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:month_picker_dialog/src/common.dart';
@@ -23,12 +22,7 @@ class MonthSelector extends StatefulWidget {
     this.firstDate,
     this.lastDate,
     this.locale,
-  })  : assert(openDate != null),
-        assert(selectedDate != null),
-        assert(onMonthSelected != null),
-        assert(upDownPageLimitPublishSubject != null),
-        assert(upDownButtonEnableStatePublishSubject != null),
-        super(key: key);
+  }) : super(key: key);
   @override
   State<StatefulWidget> createState() => MonthSelectorState();
 }
@@ -65,21 +59,28 @@ class MonthSelectorState extends State<MonthSelector> {
 
   Widget _getMonthButton(final DateTime date, final String locale) {
     final bool isEnabled = _isEnabled(date);
-    return FlatButton(
+    return ElevatedButton(
       onPressed: isEnabled
           ? () => widget.onMonthSelected(DateTime(date.year, date.month))
           : null,
-      color: date.month == widget.selectedDate!.month &&
-              date.year == widget.selectedDate!.year
-          ? Theme.of(context).accentColor
-          : null,
-      textColor: date.month == widget.selectedDate!.month &&
-              date.year == widget.selectedDate!.year
-          ? Theme.of(context).accentTextTheme.button!.color
-          : date.month == DateTime.now().month &&
-                  date.year == DateTime.now().year
-              ? Theme.of(context).accentColor
-              : null,
+      style: ButtonStyle(
+        backgroundColor: date.month == widget.selectedDate!.month &&
+                date.year == widget.selectedDate!.year
+            ? MaterialStateProperty.resolveWith(
+                (states) => Theme.of(context).colorScheme.primary)
+            : null,
+        textStyle: MaterialStateProperty.resolveWith(
+          (states) => TextStyle(
+            color: date.month == widget.selectedDate!.month &&
+                    date.year == widget.selectedDate!.year
+                ? Theme.of(context).colorScheme.onSecondary
+                : date.month == DateTime.now().month &&
+                        date.year == DateTime.now().year
+                    ? Theme.of(context).colorScheme.secondary
+                    : null,
+          ),
+        ),
+      ),
       child: Text(
         DateFormat.MMM(locale).format(date),
       ),
