@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:month_picker_dialog/src/common.dart';
@@ -22,11 +21,7 @@ class YearSelector extends StatefulWidget {
     this.firstDate,
     this.lastDate,
     this.locale,
-  })  : assert(initialDate != null),
-        assert(onYearSelected != null),
-        assert(upDownPageLimitPublishSubject != null),
-        assert(upDownButtonEnableStatePublishSubject != null),
-        super(key: key);
+  }) : super(key: key);
   @override
   State<StatefulWidget> createState() => YearSelectorState();
 }
@@ -51,7 +46,8 @@ class YearSelectorState extends State<YearSelector> {
         crossAxisCount: 4,
         children: List<Widget>.generate(
           12,
-          (final int index) => _getYearButton(page, index, getLocale(context, selectedLocale: widget.locale)),
+          (final int index) => _getYearButton(
+              page, index, getLocale(context, selectedLocale: widget.locale)),
         ).toList(growable: false),
       );
 
@@ -60,14 +56,22 @@ class YearSelectorState extends State<YearSelector> {
         page * 12 +
         index;
     final bool isEnabled = _isEnabled(year);
-    return FlatButton(
+    return ElevatedButton(
       onPressed: isEnabled ? () => widget.onYearSelected(year) : null,
-      color: year == widget.initialDate!.year
-          ? Theme.of(context).accentColor
-          : null,
-      textColor: year == widget.initialDate!.year
-          ? Theme.of(context).accentTextTheme.button!.color
-          : year == DateTime.now().year ? Theme.of(context).accentColor : null,
+      style: ButtonStyle(
+        backgroundColor: year == widget.initialDate!.year
+            ? MaterialStateProperty.resolveWith(
+                (states) => Theme.of(context).colorScheme.secondary)
+            : null,
+        textStyle: MaterialStateProperty.resolveWith(
+          (states) => TextStyle(
+              color: year == widget.initialDate!.year
+                  ? Theme.of(context).colorScheme.onSecondary
+                  : year == DateTime.now().year
+                      ? Theme.of(context).colorScheme.secondary
+                      : null),
+        ),
+      ),
       child: Text(
         DateFormat.y(locale).format(DateTime(year)),
       ),
@@ -94,7 +98,8 @@ class YearSelectorState extends State<YearSelector> {
       if (widget.lastDate!.year - widget.firstDate!.year <= 12)
         return 1;
       else
-        return ((widget.lastDate!.year - widget.firstDate!.year + 1) / 12).ceil();
+        return ((widget.lastDate!.year - widget.firstDate!.year + 1) / 12)
+            .ceil();
     } else if (widget.firstDate != null && widget.lastDate == null) {
       return (_getItemCount() / 12).ceil();
     } else if (widget.firstDate == null && widget.lastDate != null) {
@@ -119,7 +124,8 @@ class YearSelectorState extends State<YearSelector> {
     _pageController = new PageController(
         initialPage: widget.firstDate == null
             ? (widget.initialDate!.year / 12).floor()
-            : ((widget.initialDate!.year - widget.firstDate!.year) / 12).floor());
+            : ((widget.initialDate!.year - widget.firstDate!.year) / 12)
+                .floor());
     super.initState();
     new Future.delayed(Duration.zero, () {
       widget.upDownPageLimitPublishSubject.add(new UpDownPageLimit(
